+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Change Log
 ==========
 
+Unreleased
+----------
+- Support for model constraints
+
 v2.0.1
 ------
 - Remove unnecessary import of `six`